% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plg.GO.R
\name{annotateAll}
\alias{annotateAll}
\title{annotateAll}
\usage{
annotateAll(markerslist, go = NULL, topN = 10, ...)
}
\arguments{
\item{markerslist}{A list of markers table (one table per resolution), 
as produced by `Seurat::FindAllMarkers`}

\item{go}{An optional named list of genesets. If NULL, will be fetched 
using `getAllGOlists`.}

\item{topN}{The top number of terms to report (default 10)}

\item{...}{Passed to `getAllGOlists` if `go=NULL`}
}
\value{
A nested list of data.frames.
}
\description{
annotateAll
}
