% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustView.R
\name{clustView}
\alias{clustView}
\title{clustView}
\usage{
clustView(seurat, enableClustree = T)
}
\arguments{
\item{seurat}{An object of class `seurat`}

\item{enableClustree}{Logical; whether or not to enable the
clustering tree (default TRUE). Turn off to reduce loading
time.}
}
\description{
clustView
}
