% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustView.server.R
\name{clustView.server}
\alias{clustView.server}
\title{clustView.server}
\usage{
clustView.server(seurat, enableClustree = T)
}
\arguments{
\item{seurat}{An object of class `seurat`, or a path to a RDS of such an object.}

\item{enableClustree}{Logical; whether to draw the clustering tree 
(default TRUE). Disable to increase speed.}
}
\description{
Server-side function of the `clustView` shiny app.
}
