% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plg.GO.R
\name{getAllGOlists}
\alias{getAllGOlists}
\title{getAllGOlists}
\usage{
getAllGOlists(species = "Mm", translate.cat.ids = TRUE,
  gene.id.type = "symbol", genes = NULL, minSize = 10,
  maxSize = 500, returnGeneSets = FALSE, categories = NULL,
  ontologies = c("BP", "CC", "MF"))
}
\arguments{
\item{species}{Two-letter species code (corresponding to the 'org.Xx.eg' package)}

\item{translate.cat.ids}{Logical; whether to translate GO category IDs to terms}

\item{gene.id.type}{The type of gene identifier to return; 
either 'symbol' (default) or 'ensembl'.}

\item{genes}{The subset of genes to use.}

\item{minSize}{The minimum set size (default 10).}

\item{maxSize}{The maximum set size (default 500).}

\item{returnGeneSets}{Logical; whether to return an object of class 
`GeneSetCollection` rather than a normal list (default FALSE).}

\item{categories}{The GO categories to use (default all)}

\item{ontologies}{The GO tree to use; any combination of "BP", "CC", and "MF" 
(default all)}
}
\value{
A named list of genes, or a `GeneSetCollection` if `returnGeneSets=T`
}
\description{
Returned as list of genes for each GO term
}
