% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plg.GO.R
\name{prepSeuratForClustView}
\alias{prepSeuratForClustView}
\title{prepSeuratForClustView}
\usage{
prepSeuratForClustView(seurat, markerslist, res_prefixes = NULL,
  species = "Mm", ontologies = c("BP", "CC"))
}
\arguments{
\item{seurat}{An object of class `Seurat`}

\item{markerslist}{A nested list of markers' tables (as produced by 
`Seurat::FindAllMarkers()`) for each method and resolution}

\item{res_prefixes}{The list of clustering prefixes to use from 
`seurat@meta.data` (default all detected)}

\item{species}{Two-letter species code (corresponding to the 
'org.Xx.eg' package), default 'Mm'.}

\item{ontologies}{The GO tree to use; any combination of 
"BP", "CC", and "MF" (default `c("BP", "CC")`)}
}
\value{
The `seurat` object, with cluster annotation in the slot
`seurat@misc$clusterAnnotation`
}
\description{
prepSeuratForClustView
}
